+++ app/models/carto/user.rb
@@ -93,6 +93,10 @@ class Carto::User < ActiveRecord::Base
   after_destroy { rate_limit.destroy_completely(self) if rate_limit }
   after_destroy :invalidate_varnish_cache
 
+  def sequel_user
+    ::User.find(id: id)
+  end
+
   # Auto creates notifications on first access
   def static_notifications_with_creation
     static_notifications_without_creation || build_static_notifications(user: self, notifications: {})